🗃Refactor: template.ts 삭제 및 createElement 함수 이동
